Allow for custom total counter in pagination
This is related to issue #518, if you have some really large tables and
want to provide a total count you may want to tweak how the total gets
calculated. The default query is slower than just doing a COUNT(*), but
you might also want to cache the response in Redis or something similar
between calls.

So pass in a function or lambda that takes the current class and do
whatever you need and return an integer back.